Updated with-sitemap example for App Router (#66995)
Hello

This PR updates the with-sitemap example to use:

1. App Router
2. TypeScript
3. sitemap.js

---------
