+++ examples/with-sitemap/app/layout.tsx
@@ -1,0 +1,19 @@
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by Next.js",
+  icons: {
+    icon: "/favicon.ico",
+  },
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body>{children}</body>
+    </html>
+  );
+}
