+++ examples/with-sitemap/next.config.js
@@ -1,10 +1,2 @@
 /** @type {import('next').NextConfig} */
-module.exports = {
-  webpack: (config, { isServer }) => {
-    if (isServer) {
-      require("./scripts/generate-sitemap");
-    }
-
-    return config;
-  },
-};
+module.exports = {};
