+++ examples/with-sitemap/package.json
@@ -11,6 +11,10 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "globby": "^11.0.1"
+    "globby": "^11.0.1",
+    "@types/node": "^18.11.5",
+    "@types/react": "^18.0.23",
+    "@types/react-dom": "^18.0.7",
+    "typescript": "^4.8.4"
   }
 }
