--- examples/with-sitemap/public/sitemap.xml
@@ -1,10 +1,0 @@
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-  <url>
-    <loc>http://localhost:3000/contact</loc>
-    <changefreq>hourly</changefreq>
-  </url>
-  <url>
-    <loc>http://localhost:3000</loc>
-    <changefreq>hourly</changefreq>
-  </url>
-</urlset>
